move fastlanes modules and files around
Also renames `compress` to `bitpack_compress` since it gets exported out
of `fastlanes` and you have no idea which method compress actually
means.
