+++ encodings/alp/src/alp_rd/mod.rs
@@ -7,6 +7,7 @@ pub use array::*;
 use vortex_array::IntoArray;
 use vortex_array::patches::Patches;
 use vortex_array::validity::Validity;
+use vortex_fastlanes::bitpack_compress::bitpack_encode_unchecked;
 
 mod array;
 mod compute;
@@ -24,7 +25,6 @@ use vortex_array::{Array, ToCanonical};
 use vortex_buffer::{Buffer, BufferMut};
 use vortex_dtype::{DType, NativePType, match_each_integer_ptype};
 use vortex_error::{VortexExpect, VortexUnwrap, vortex_panic};
-use vortex_fastlanes::bitpack_encode_unchecked;
 use vortex_utils::aliases::hash_map::HashMap;
 
 use crate::match_each_alp_float_ptype;
