+++ encodings/fastlanes/benches/pipeline_bitpacking_compare_scalar.rs
@@ -14,7 +14,8 @@ use vortex_buffer::{BitBuffer, BufferMut};
 use vortex_dtype::NativePType;
 use vortex_error::VortexResult;
 use vortex_expr::{Scope, lit, lt, root};
-use vortex_fastlanes::{FoRArray, bitpack_to_best_bit_width};
+use vortex_fastlanes::FoRArray;
+use vortex_fastlanes::bitpack_compress::bitpack_to_best_bit_width;
 use vortex_mask::Mask;
 use vortex_scalar::Scalar;
