+++ encodings/fastlanes/src/bitpacking/compute/mod.rs
@@ -45,8 +45,9 @@ mod tests {
     use vortex_array::compute::conformance::binary_numeric::test_binary_numeric_array;
     use vortex_array::compute::conformance::consistency::test_array_consistency;
 
+    use crate::BitPackedArray;
+    use crate::bitpack_compress::bitpack_encode;
     use crate::bitpacking::compute::chunked_indices;
-    use crate::{BitPackedArray, bitpack_encode};
 
     #[test]
     fn chunk_indices_repeated() {
