+++ encodings/fastlanes/src/bitpacking/vtable/visitor.rs
@@ -1,0 +1,20 @@
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-FileCopyrightText: Copyright the Vortex contributors
+
+use vortex_array::vtable::{ValidityHelper, VisitorVTable};
+use vortex_array::{ArrayBufferVisitor, ArrayChildVisitor};
+
+use crate::{BitPackedArray, BitPackedVTable};
+
+impl VisitorVTable<BitPackedVTable> for BitPackedVTable {
+    fn visit_buffers(array: &BitPackedArray, visitor: &mut dyn ArrayBufferVisitor) {
+        visitor.visit_buffer(array.packed());
+    }
+
+    fn visit_children(array: &BitPackedArray, visitor: &mut dyn ArrayChildVisitor) {
+        if let Some(patches) = array.patches() {
+            visitor.visit_patches(patches);
+        }
+        visitor.visit_validity(array.validity(), array.len());
+    }
+}
